mv: CgnScreenStatus to own file
